Check in work so far on schema object
