+++ siliconcompiler/schema/__init__.py
@@ -1,0 +1,2 @@
+from .schema_cfg import SCHEMA_VERSION
+from .schema_obj import Schema
